refactor: add new `Event`, `EventType` and `EventParam`
